cride/requirements/base.txt
# Base
pytz==2023.3.post1
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.28

# Environment
django-environ==0.4.5

# DjangoRestFramework
djangorestframework==3.9.3

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.4.7
tornado>=5.0.0,<6.0.0
flower==0.9.2